Refactored purchases a bit
- Changed PurchaseRequest so userId is only asked once
- Added validation to the new PurchaseRequest
- Updated documentation to reflect new changes
