+++ ticketguru/src/main/java/copypaste/ticketguru/domain/PurchaseRequest.java
@@ -2,29 +2,18 @@ package copypaste.ticketguru.domain;
 
 import java.util.List;
 
-import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 
 public class PurchaseRequest {
-    @NotNull(message = "Event cannot be NULL")
-	private Long eventId;
-
     @NotNull(message = "User cannot be NULL")
     private Long userId;
 
-    @NotEmpty(message = "Tickets cannot be empty")
-    @NotNull(message = "Tickets cannot be NULL")
-    private List<@NotBlank String> ticketTypeNames;
     // Would be good to also have validation for ticketTypes being in the valid set, but that'd require stupid code
-
-    public Long getEventId() {
-        return eventId;
-    }
-
-    public void setEventId(Long eventId) {
-        this.eventId = eventId;
-    }
+    @NotEmpty(message = "Request rows cannot be empty")
+    @NotNull(message = "Request rows cannot be NULL")
+    private List<@Valid PurchaseRequestRow> purchaseRequestRows;
 
     public Long getUserId() {
         return userId;
@@ -34,11 +23,11 @@ public class PurchaseRequest {
         this.userId = userId;
     }
 
-    public List<String> getTicketTypeNames() {
-        return ticketTypeNames;
+    public List<PurchaseRequestRow> getPurchaseRequestRows() {
+        return this.purchaseRequestRows;
     }
 
-    public void setTicketTypeNames(List<String> ticketTypeNames) {
-        this.ticketTypeNames = ticketTypeNames;
+    public void setPurchaseRequestRows(List<PurchaseRequestRow> rows) {
+        this.purchaseRequestRows = rows;
     }
 }
