+++ ticketguru/src/main/java/copypaste/ticketguru/web/rest/PurchaseRestController.java
@@ -14,19 +14,23 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import copypaste.ticketguru.domain.AppUser;
 import copypaste.ticketguru.domain.Purchase;
 
 import copypaste.ticketguru.domain.PurchaseRepository;
 import copypaste.ticketguru.domain.PurchaseRequest;
+import copypaste.ticketguru.domain.UserRepository;
 import copypaste.ticketguru.service.PurchaseService;
+import jakarta.validation.Valid;
 
 @RestController
 public class PurchaseRestController {
-
 	@Autowired
     private PurchaseService purchaseService;
 	@Autowired
 	private PurchaseRepository purchaseRepository;
+    @Autowired
+    private UserRepository userRepository;
 
 	@GetMapping(value = "/api/purchases")
 	public List<Purchase> getAllPurchases() {
@@ -36,13 +40,21 @@ public class PurchaseRestController {
 	// Luodaan ostotapahtuma, jossa voidaan ostaa lippuja useampaan tapahtumaan
 	@PostMapping("/api/purchases")
     @Transactional
-    public ResponseEntity<?> createMultiplePurchasesWithTickets(@RequestBody List<PurchaseRequest> purchaseRequests) {
+    public ResponseEntity<?> createMultiplePurchasesWithTickets(@Valid @RequestBody PurchaseRequest purchaseRequest) {
         List<Purchase> successfulPurchases = new ArrayList<>();
         List<String> failedPurchases = new ArrayList<>();
 
+        Optional<AppUser> userOpt = userRepository.findById(purchaseRequest.getUserId());
+
+        if(userOpt.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+
+        AppUser user = userOpt.get();
+
 		// Käsittelee jokaisen tapahtuman osto itsenäisesti
-        purchaseRequests.forEach(request -> {
-            Optional<Purchase> result = purchaseService.processPurchaseRequest(request);
+        purchaseRequest.getPurchaseRequestRows().forEach(request -> {
+            Optional<Purchase> result = purchaseService.processPurchaseRequest(request, user);
             result.ifPresentOrElse(
                 successfulPurchases::add,
                 () -> failedPurchases.add("Ostotapahtumaa ei pystytty suorittamaan tapahtumaan ID:llä: " + request.getEventId() + ". Tarkasta tapahtuman ID, lipputyypit ja saatavuus.")
